implement different serializers
This commit reviews the serialization approach from the ground to allow using
different serializers and deserializers. For instance, this allows avoiding a
copy after calling `.serialize()`: in the previous version, `serialize()` was
always returning a `Zeroizing<Vec<u8>>`; in this new version, `serialize` takes
an additional type parameter specifying which `Serializer` to use and returns
`Serializer::Output`.

Two serializers are implemented:

- `SecretSerializer` which returns a `Zeroizing<Vec<u8>>`; calling
`.serialize<SecretSerializer>()` is therefore equivalent to the previous
`.serialze()` call.

- `PublicSerializer` which returns a `Vec<u8>` and is therefore appropriate to
for use in deserialization of values that do not need to be zeroized on drop.